converted buffer tests to typescript
